Extract energy gas graph card series generation into a pure module
